# Conflicts:
#	client-interface/src/main/java/com/epam/clientinterface/controller/SignUpController.java
#	client-interface/src/main/java/com/epam/clientinterface/controller/advice/ErrorHandlingAdvice.java
#	client-interface/src/main/resources/liquibase/changelog-master.xml
#	client-interface/src/test/java/com/epam/clientinterface/controller/SignUpControllerTest.java
#	client-interface/src/test/java/com/epam/clientinterface/service/AuthServiceImplTest.java

+++ client-interface/src/main/java/com/epam/clientinterface/controller/dto/response/ErrorResponse.java
@@ -7,10 +7,10 @@ import org.springframework.http.HttpStatus;
 @Getter
 public class ErrorResponse {
     private final String type;
-    private final short status;
+    private final int status;
 
     public ErrorResponse(@NonNull String type, @NonNull HttpStatus status) {
         this.type = type;
-        this.status = (short) status.value();
+        this.status = status.value();
     }
 }
